analyzer/src/funTest/kotlin/PubTest.kt
/*
 * Copyright (C) 2017-2019 HERE Europe B.V.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 * License-Filename: LICENSE
 */

package com.here.ort.analyzer

import com.here.ort.analyzer.managers.Pub
import com.here.ort.downloader.VersionControlSystem
import com.here.ort.model.yamlMapper
import com.here.ort.utils.normalizeVcsUrl
import com.here.ort.utils.test.DEFAULT_ANALYZER_CONFIGURATION
import com.here.ort.utils.test.DEFAULT_REPOSITORY_CONFIGURATION
import com.here.ort.utils.test.USER_DIR
import com.here.ort.utils.test.patchExpectedResult

import io.kotlintest.shouldBe
import io.kotlintest.specs.WordSpec

import java.io.File

class PubTest : WordSpec() {
    private val projectsDir = File("src/funTest/assets/projects/synthetic/pub/").absoluteFile
    private val vcsDir = VersionControlSystem.forDirectory(projectsDir)!!
    private val vcsUrl = vcsDir.getRemoteUrl()
    private val vcsRevision = vcsDir.getRevision()

    init {

        "Pub" should {
            "resolve dart http dependencies correctly" {
                val workingDir = File(projectsDir, "http")
                val packageFile = File(workingDir, "pubspec.yaml")
                var expectedResultFile = File(projectsDir.parentFile, "pub-expected-output.yml");

                val result = createPub().resolveDependencies(listOf(packageFile))[packageFile]
                val resultString = yamlMapper.writeValueAsString(result);

                expectedResultFile.writeText(resultString);


                /*val vcsPath = vcsDir.getPathToRoot(workingDir)
                val expectedResult = patchExpectedResult(
                    expectedResultFile,
                    custom = Pair("pub-project", "pub-${workingDir.name}"),
                    definitionFilePath = "$vcsPath/pubspec.yaml",
                    url = normalizeVcsUrl(vcsUrl),
                    revision = vcsRevision,
                    path = vcsPath
                )

                yamlMapper.writeValueAsString(result) shouldBe expectedResult*/
            }
/*
            "resolve package-lock dependencies correctly" {
                val workingDir = File(projectsDir, "package-lock")
                val packageFile = File(workingDir, "pubspec.yaml")
                var expectedResultFile = File(projectsDir.parentFile, "pub-expected-output.yml");

                val result = createPub().resolveDependencies(listOf(packageFile))[packageFile]
                val resultString = yamlMapper.writeValueAsString(result);

                print("resultString: $resultString");

                expectedResultFile.writeText(resultString);


                val vcsPath = vcsDir.getPathToRoot(workingDir)
                val expectedResult = patchExpectedResult(
                    expectedResultFile,
                    custom = Pair("pub-project", "pub-${workingDir.name}"),
                    definitionFilePath = "$vcsPath/pubspec.yaml",
                    url = normalizeVcsUrl(vcsUrl),
                    revision = vcsRevision,
                    path = vcsPath
                )

                yamlMapper.writeValueAsString(result) shouldBe expectedResult
            }
*/
            /*"resolve no-lockfile dependencies correctly" {
                val workingDir = File(projectsDir, "no-lockfile")
                val packageFile = File(workingDir, "pubspec.yaml")
                var expectedResultFile = File(projectsDir.parentFile, "pub-expected-output.yml");

                val result = createPub().resolveDependencies(listOf(packageFile))[packageFile]
                val resultString = yamlMapper.writeValueAsString(result);

                print("resultString: $resultString");

                expectedResultFile.writeText(resultString);


                val vcsPath = vcsDir.getPathToRoot(workingDir)
                val expectedResult = patchExpectedResult(
                    expectedResultFile,
                    custom = Pair("pub-project", "pub-${workingDir.name}"),
                    definitionFilePath = "$vcsPath/pubspec.yaml",
                    url = normalizeVcsUrl(vcsUrl),
                    revision = vcsRevision,
                    path = vcsPath
                )

                yamlMapper.writeValueAsString(result) shouldBe expectedResult
            }*/
        }
    }

    private fun createPub() =
        Pub("Pup", USER_DIR, DEFAULT_ANALYZER_CONFIGURATION, DEFAULT_REPOSITORY_CONFIGURATION)
}
